requirements.txt
# Framework web
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Bootstrap-Flask==2.3.0
python-dotenv==1.0.0

# Base de datos
SQLAlchemy==2.0.23
PyMySQL==1.1.0

# Generación y lectura de QR
qrcode==7.4.2
opencv-python==4.8.1.78
pyzbar==0.1.9

# Utilidades
python-dateutil==2.8.2
Werkzeug==3.0.1

# Para generar PDFs (tickets/comprobantes)
reportlab==4.0.7
# Impresora térmica ESC/POS
python-escpos==3.0.0
# Para desarrollo
pytest==7.4.3
black==23.10.1
flake8==6.1.0

